pyproject.toml
[build-system]
requires = ['setuptools']
build-backend = 'setuptools.build_meta'

[project]
name = 'mibiao'
version = '0.0.1'
requires-python = '>=3.12,<3.13'
dependencies = [
    'click==8.1.7',
    'Werkzeug==3.1.3',
    'Flask==3.1.0',
    'email_validator==2.2.0',
    'WTForms==3.2.1',
    'Flask-WTF==1.2.2',
    'SQLAlchemy==2.0.36',
    'SQLAlchemy-Utils==0.41.2',
    'Flask-SQLAlchemy==3.1.1',
    'Flask-Login==0.6.3',
    'Flask-Compress==1.17.0',
    'pendulum==3.0.0',
    'python-dotenv==1.0.1',
    'gevent==24.11.1',
    'gunicorn==23.0.0',
    'filelock==3.16.1',
    'passlib[bcrypt]==1.7.4',
    'bcrypt==4.2.1',
    'shortuuid==1.0.13',
    'mistune==3.0.2',
    'openpyxl==3.1.5',
    'psutil==6.1.0',
    'memory-profiler==0.61.0',
]
[project.scripts]
mibiao = 'mibiao.cli:cli'

[tool.setuptools.packages.find] # https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html
where = ['src']
include = ['*']

[tool.setuptools.package-data]
'*' = ['*.*']

[tool.black]
target-version = ["py312"]
line-length = 120

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
ignore = ["E402"]

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint.isort]
force-single-line = true
case-sensitive = true
order-by-type = true

[tool.ruff.lint.extend-per-file-ignores]
'*' = ['E402']

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"
